changed file directories for a more mvc approach
