+++ server/Storage/index.js
@@ -1,19 +1,7 @@
-const express = require("express")
-const stream = require("stream")
 
-const app = express()
+// Imports: The log and date_diff_in_minute utility functions are imported from an external module.
+const {log, date_diff_in_minute} = require('../utils');
 
-const PORT = Number(process.env.PORT) || 4000
-
-app.use(express.json());
-app.use(express.urlencoded({extended: true}));
-
-/**
- * Utils
- * @param messages
- */
-const log = (...messages) => console.log(new Date(), ...messages)
-const date_diff_in_minute = (date1, date2) => Math.abs(Math.floor((date1 - date2)/60_000));
 
 /**
  * Storage System
@@ -93,66 +81,7 @@ const STORAGE = new (class {
         log("IN MEMORY REPORT", String(Object.keys(this.data)))
     }
 
-})()
+});
 
-app.get('/', (req, res) => {
-    const {id, pwd} = STORAGE.request_slot()
-    res.status(200).json({id, pwd, message: "Success"})
-})
-
-app.get('/:id', (req, res) => {
-    const id = req.params.id
-    const data = STORAGE.read(id)
-
-    if (data === null) {
-        res.status(404).json({
-            message: "Not Found"
-        })
-        return
-    }
-
-    res.status(200).json({
-        data,
-        message: "Success"
-    })
-
-})
-
-app.get('/file/:id', (req, res) => {
-    const id = req.params.id
-    const data = STORAGE.read(id)
-
-    if (data === null) {
-        res.status(404).json({
-            message: "Not Found"
-        })
-        return
-    }
-
-    const fileContents = Buffer.from(data, "utf-8");
-    const readStream = new stream.PassThrough();
-    readStream.end(fileContents);
-
-    res.set('Content-disposition', `attachment; filename=Insomnia-${id}.json`);
-    res.set('Content-Type', 'application/json');
-
-    readStream.pipe(res);
-})
-
-app.post('/:id', (req, res) => {
-    const {data} = req.body
-    const {id} = req.params
-    const {pwd} = req.query
-    if (id && pwd && data) {
-        if (STORAGE.save(id, pwd, data)) {
-            res.status(200).json({message: "Success"})
-            return
-        }
-        res.status(400).json({message: "Incorrect password (pwd)"})
-        return
-    }
-    res.status(400).json({message: "Bad request details are missing"})
-})
-
-// setInterval(STORAGE.clean_older_slots, 10*60_000)
-app.listen(PORT, () => log("STARTED API GATEWAY AT", `PORT:${PORT}`))
+// Exports: The instantiated STORAGE object is exported, which means other modules can require and use this object.
+module.exports = STORAGE;
